pyproject.toml
[project]
name = "shotcheck"
version = "0.1.0"
description = "A screenshot comparison tool for Visual Regression Testing"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "pillow>=12.1.1",
]
